% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bb_gecco.R
\docType{package}
\name{bb_gecco}
\alias{bb_gecco}
\alias{bb_gecco-package}
\alias{gecco}
\title{bb_gecco}
\description{
bb_gecco: Barbs functions for fetching gecco and isacc data
}
\details{
\tabular{ll}{
Package: \tab bb_gecco \cr
Type: \tab Package \cr
Version: \tab 1.0 \cr
Date: \tab 2015-05-22 \cr
License: \tab GPL(>= 2) \cr
}
}
\author{
Barb Banbury
}

