% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{ChangeAlleles}
\alias{ChangeAlleles}
\title{Change Alleles from Numeric to Letters}
\usage{
ChangeAlleles(num)
}
\arguments{
\item{num}{A number 1:4}
}
\description{
Change Allele Coding from 1,2,3,4 to A,C,T,G
}
\details{
This function changes numeric allele designation to letters
}
\examples{
ChangeAlleles("1")
}
\seealso{
\link{MakePriorityPrunerInputFile}
}

