% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{ChangeStudyNames}
\alias{ChangeStudyNames}
\title{Change Study Names}
\usage{
ChangeStudyNames(study)
}
\arguments{
\item{study}{One of: ASTERISK, OFCCR, PMH-CCFR}
}
\value{
Returns a new study name
}
\description{
This function will change the study names from those in the ISACC paper proposal to the numerical name.
}
\examples{
ChangeStudyNames("ASTERISK")
sapply(c("ASTERISK", "OFCCR", "PMH-CCFR"), ChangeStudyNames)
}

