% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{CreateDosageDataAcrossStudies}
\alias{CreateDosageDataAcrossStudies}
\title{Create a Dosage Dataset from HapMap}
\usage{
CreateDosageDataAcrossStudies(rs_positions, saveToRdata = NULL)
}
\arguments{
\item{rs_positions}{An object of the class "snp_location_info" that comes out of the FindSNPpositions* functions}

\item{saveToRdata}{Optional file name to save Rdata to. Will not save if left NULL.}
}
\description{
Create dataset with Dosage Information for one or many hapmap studies
}
\details{
This function uses the information in rs_positions (class snp_location_info) to gather a
dosage dataset.  There can be one study or many studies. Returns a dataset with individuals
as rows and rs_positions as columns.
}
\examples{
tmp <- FindSNPposition_hapmap(c("rs2736100", "rs401681", "rs10069690"), studies="101ccfr")
CreateDosageDataAcrossStudies(tmp)
}
\seealso{
\link{FindSNPposition_hapmap} \link{CreateDosageDataPerStudy}
}

