% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{CreateDosageDataFromGigs}
\alias{CreateDosageDataFromGigs}
\title{Create Dosage Dataset in GIGSv2}
\usage{
CreateDosageDataFromGigs(gigs_positions, chatty = TRUE,
  report.times = FALSE)
}
\arguments{
\item{gigs_positions}{An object of the class "snp_location_info" that comes out of the FindSNPpositions* functions}

\item{chatty}{Option to print progress to screen}

\item{report.times}{Option to return computational time with dataset as a second object in the list}
}
\description{
Create a new Dosage Dataset from GIGSv2
}
\details{
This function will create a gigs dosage dataset, with individuals as rows and snps as columns. If times
is selected, then the output wll be a list where the first element is the dosage data, and the second
element is the computational time.
}
\examples{
tmp <- FindSNPpositions_gigs(c("5:1286516", "5:1322087", "5:1279790"))
CreateDosageDataFromGigs(tmp)
}
\seealso{
\link{CreateDosageDataAcrossStudies} \link{MakeSNPDetailsTable_GIGS} \link{Use_snp_finder.py}
}

