% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{CreateDosageDataPerStudy}
\alias{CreateDosageDataPerStudy}
\title{Create a Dosage Dataset One Study}
\usage{
CreateDosageDataPerStudy(rs_positions, directory = NULL)
}
\arguments{
\item{rs_positions}{An object of the class "snp_location_info" that comes out of the FindSNPpositions* functions}

\item{directory}{File path addition to find the nc files}
}
\description{
Create dataset with Dosage Information for one hapmap study
}
\details{
Somewhat of an internal function that returns a dataset with individuals as rows and rs_positions as columns for just one hapmap study (ex,"101ccfr" will return both "usc" and "usc2" batches). Internal, because it is hard to give the directory. If you don't want to have to figure that out, then use the function \link{CreateDosageDataAcrossStudies}
}
\examples{
tmp <- FindSNPposition_hapmap(c("rs2736100", "rs401681", "rs10069690"), studies="102arctic")
dir <- "102arctic/102arctic_minusccfr/mach/"
CreateDosageDataPerStudy(tmp, directory=paste0(DataLocation("hapmap"), dir))
}
\seealso{
\link{FindSNPposition_hapmap} \link{CreateDosageDataAcrossStudies}
}

