% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{CreateEpiDataset}
\alias{CreateEpiDataset}
\title{CreateEpiDataset}
\usage{
CreateEpiDataset(variables, studies = NULL, files = "1or2", chatty = TRUE)
}
\arguments{
\item{variables}{Which variables to collect (ex: "smoke", "BMI")}

\item{studies}{Which studies to pull, if left NULL it will use all}

\item{files}{Which files to use from the study (this is bad...)}

\item{chatty}{Option to print progress to screen}
}
\value{
Returns a dataset of subjects as rows, and compassID, netcdfID, study name, and variables as columns
}
\description{
Create an Epi dataset
}
\details{
This function will return a table with subjects as rows and variables as columns. It gets data
from a single or multiple stuies.
}
\examples{
CreateEpiDataset("smoke", "101ccfr")
CreateEpiDataset(c("smoke", "BMI"))
}
\seealso{
\link{CreateEpiDatasetPerStudy} \link{Yi_GetEpiDataFromGigs}
}

