% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{CreateEpiDatasetPerStudy}
\alias{CreateEpiDatasetPerStudy}
\title{CreateEpiDatasetPerStudy}
\usage{
CreateEpiDatasetPerStudy(variables, study, files = "1or2", chatty = TRUE)
}
\arguments{
\item{variables}{Which variables to collect (ex: "smoke", "BMI")}

\item{study}{Which study data to pull}

\item{files}{Which files to use from the study (this is bad...)}

\item{chatty}{Option to print progress to screen}
}
\value{
Returns a dataset of subjects x variables
}
\description{
Create an Epi dataset from just a single study
}
\details{
This function will return a table with subjects as rows and variables as columns. It only gets data
from a single study, so in this way it is really an internal function for \link{CreateEpiDataset}.
}
\examples{
CreateEpiDatasetPerStudy("smoke", "101ccfr")
}
\seealso{
\link{CreateEpiDataset}
}

