% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{CreateSurvivalDataset}
\alias{CreateSurvivalDataset}
\title{CreateSurvivalDataset}
\usage{
CreateSurvivalDataset(variables = "all", studies = NULL,
  data = "pooledGecco")
}
\arguments{
\item{variables}{Which variables to collect. If "all", then it returns all columns with data.}

\item{data}{Which data files to use. There are several different survival datasets floating about, you can choose to use the ones for pooled GECCO, separate GECCO, or separate ISACC}

\item{study}{Which studies to pull, if NULL, then it returns all studies}

\item{chatty}{Option to print progress to screen}
}
\value{
Returns a dataset of subjects as rows, and compassID, netcdfID, study name, and variables as columns
}
\description{
Create a Survival dataset
}
\details{
This function will return a table with subjects as rows and variables as columns. It gets data
from a single or multiple stuies.
}
\examples{
CreateSurvivalDataset("all")
l <- CreateSurvivalDataset("censor", data="separateGecco")
}
\seealso{
\link{CreateEpiDatasetPerStudy}
}

