% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{DataLocation}
\alias{DataLocation}
\title{List of file directories for gecco data}
\usage{
DataLocation(whichData)
}
\arguments{
\item{whichData}{Specifies which data you need, and will point to a file/directory}
}
\value{
Returns a path to the directory. THis function is where we can store where data lives in directories,
so that when we change where data lives, we only have to change it once.
}
\description{
This function will add the specific path plus root directory
}
\examples{
DataLocation("hapmap_snp_names")
}
\seealso{
\link{MakePathtoNewcomb_P} \link{MakePathtoPeters_U}
}

