% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{deFactorize}
\alias{deFactorize}
\title{deFactorize}
\usage{
deFactorize(table)
}
\arguments{
\item{table}{Any table (class data.frame or matrix)}
}
\description{
Auto Remove the factoring
}
\details{
This function changes any factor class to character
}

