% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{Find_position_gigs_single_chromo}
\alias{Find_position_gigs_single_chromo}
\title{Find position in gigsv2}
\usage{
Find_position_gigs_single_chromo(snp_name, chromosome)
}
\arguments{
\item{snp_name}{A vector of snp names in gigs format (ie, 1:1234)}

\item{chromosome}{which chromosome to search. Could be written smarter so that it auto finds this from the position name.  Duh.}
}
\description{
Find a snp position in the gigsv2 dataset
}
\details{
This function will use a gigs snp name (ie, 2:123) and find where that is located in the gigs dataset.
This function only searches a single chromosome, so can also use \link{FindSNPpositions_gigs} if you
need to do more than one. Returns a matrix with study (always gigs), gene(always NA), the snp name,
the ncdf file it is located, and the position within the file.
}
\examples{
Find_position_gigs_single_chromo(c("5:1286516", "5:1322087", "5:1279790"), 5)
}
\seealso{
\link{FindSNPpositions_gigs}
}

