% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{find_rs_hapmap}
\alias{find_rs_hapmap}
\title{Find RS Numbers from HapMap}
\usage{
find_rs_hapmap(rs_number, snp_names)
}
\arguments{
\item{rs_number}{Vector or single SNP name}

\item{snp_names}{Which hapmap study file you want to use, these are found in the directory DataLocation("hapmap_snp_names").}
}
\description{
Find RS numbers from HapMap study-specific dataset
}
\details{
Returns table with study, batch, rs_number, ncdf file, and the position
}
\examples{
load(paste0(DataLocation("hapmap_snp_names"), "snp_names_101ccfr_usc.Rdata"))
find_rs_hapmap(c("rs2736100", "rs401681", "rs10069690"), snp_names)
}
\seealso{
\link{FindSNPposition_hapmap}
}

