% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{find_rs_to_gigs}
\alias{find_rs_to_gigs}
\title{Find RS numbers in GIGSv1 Data}
\usage{
find_rs_to_gigs(rs_number, chatty = TRUE)
}
\arguments{
\item{rs_number}{A vector of rs numbers}

\item{chatty}{Option to print progress to screen}
}
\description{
Find RS numbers in the GIGSv1 dataset (not available for GIGSv2 yet)
}
\details{
This function allows you to search an RS number for its corresponding GIGs name. Will return a table with two columns, rs number, and gigs number.
}
\examples{
find_rs_to_gigs(c("rs2736100", "rs401681", "rs10069690"))
}
\seealso{
\link{MakePriorityPrunerInputFile}
}

