% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{FindSNPposition_hapmap}
\alias{FindSNPposition_hapmap}
\title{Find RS Numbers from HapMap}
\usage{
FindSNPposition_hapmap(rs_number, studies = NULL, chatty = TRUE)
}
\arguments{
\item{rs_number}{Vector or single SNP name}

\item{studies}{Which hapmap study files you want to include, these are found in the directory DataLocation("hapmap_snp_names"). If NULL, it will include all.}

\item{chatty}{Option to print progress to screen}
}
\description{
Find RS numbers from HapMap accross various studies
}
\details{
Returns table in the class "snp_location_info", with study, batch, rs_number, ncdf file, and the position. The special class makes it so that it does not print all results to the screen unless you explicitly want to.  Use [] to display the whole table.
}
\examples{
FindSNPposition_hapmap(c("rs2736100", "rs401681", "rs10069690"), studies="101ccfr")
FindSNPposition_hapmap(c("rs2736100", "rs401681", "rs10069690"))
}
\seealso{
\link{find_rs_hapmap}
}

