% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{FindSNPpositions_gigs}
\alias{FindSNPpositions_gigs}
\title{Find position in gigsv2}
\usage{
FindSNPpositions_gigs(snp_name, chatty = TRUE)
}
\arguments{
\item{snp_name}{A vector of snp names in gigs format (ie, 1:1234) or a data frame in the class "snps_in_gene_regions"}

\item{chatty}{Option to print progress to screen}
}
\description{
Find a snp position in the gigsv2 dataset
}
\details{
This function will use a gigs snp name (ie, 2:123) and find where that is located in the gigs dataset.
This function only searches over a single or many chromosomes. Returns a matrix with study (always gigs
for gigs data), gene (if available), the snp name, the ncdf file it is located, and the position
within the file. If the snp_name you pass has gene information, then this will be preserved in the
output.  Otherwise it will be NA.
}
\examples{
FindSNPpositions_gigs(c("5:1286516", "5:1322087", "5:1279790"))
}
\seealso{
\link{Find_position_gigs_single_chromo} \link{MakeSNPDetailsTable_GIGS} \link{Use_snp_finder.py}
}

