% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{GetBatch}
\alias{GetBatch}
\title{Get Batch}
\usage{
GetBatch(ncdfFileName)
}
\arguments{
\item{ncdfFileName}{An ncdf file name with batch name number}
}
\description{
Get Batch out of ncdf file name
}
\details{
This function returns a batch from a filename
}
\examples{
GetBatch("101ccfr_usc2_merged_dosage_5.nc")
}
\seealso{
\link{GetChromo} \link{GetStudy} \link{GetLastFileNameInPath}
}

