% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{GetChromo}
\alias{GetChromo}
\title{Get Chromosome}
\usage{
GetChromo(ncdfFileName)
}
\arguments{
\item{ncdfFileName}{An ncdf file name with chromosome number at the end}
}
\value{
This function returns a chromosome from a filename
}
\description{
Get Chromosome Number
}
\examples{
GetChromo("101ccfr_usc2_merged_dosage_5.nc")
}
\seealso{
\link{GetStudy} \link{GetBatch} \link{GetLastFileNameInPath}
}

