% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{GetCountAndBaselineAlleles}
\alias{GetCountAndBaselineAlleles}
\title{GetCountAndBaselineAlleles}
\usage{
GetCountAndBaselineAlleles(rs_numbers)
}
\arguments{
\item{rs_numbers}{A vector of rs numbers}
}
\description{
Get Count and Baseline Allele Information
}
\details{
This function pulls the count and baseline allele information per locus from the legend file. Returns a
matrix with rs number, position, allele 1 (count), and allele 2 (baseline)
}
\examples{
GetCountAndBaselineAlleles(c("rs2736100", "rs401681", "rs10069690"))
}
\seealso{
\link{CreateSNPDetailsTable} \link{CreateDosageDataPerStudy}
}

