% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{GetLastFileNameInPath}
\alias{GetLastFileNameInPath}
\title{Get File Name without Path}
\usage{
GetLastFileNameInPath(fileWithPath)
}
\arguments{
\item{fileWithPath}{An ncdf file name with or without a whole path}
}
\description{
Get a filename without the whole path
}
\details{
This function strips all path information from a file, and returns just the name of the file.
}
\examples{
GetLastFileNameInPath("/directory/to/remove/101ccfr_usc2_merged_dosage_5.nc")
}
\seealso{
\link{GetChromo} \link{GetStudy} \link{GetLastFileNameInPath}
}

