% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{GetMarginal}
\alias{GetMarginal}
\title{GetMarginal Results}
\usage{
GetMarginal(snps, datasource = "gigsv2", chatty = TRUE)
}
\arguments{
\item{snps}{A gigs SNP name, chr:pos (ex, "1:1234")}

\item{datasource}{Either "gigsv1" or "gigsv2" for now, can add hapmap later}

\item{chatty}{Option to print progress to screen}
}
\description{
Get Marginal Analysis results
}
\details{
Returns a table of results, with snp as rows
}
\examples{
GetMarginal(c("2:204570092", "3:30680370", "1:169602109"))
}
\seealso{
\link{MakePriorityPrunerInputFile}
}

