% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{GetStudy}
\alias{GetStudy}
\title{Get Study}
\usage{
GetStudy(ncdfFileName)
}
\arguments{
\item{ncdfFileName}{An ncdf file name with study name number at the beginning}
}
\value{
This function returns a study from a filename
}
\description{
Get Study out of ncdf file name
}
\examples{
GetStudy("101ccfr_usc2_merged_dosage_5.nc")
}
\seealso{
\link{GetChromo} \link{GetBatch} \link{GetLastFileNameInPath}
}

