% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{GiveStudiesNumbers}
\alias{GiveStudiesNumbers}
\title{GiveStudiesNumbers}
\usage{
GiveStudiesNumbers(study)
}
\arguments{
\item{study}{A single study (ex: "whi", or "nhs")}
}
\value{
Will return a new study name
}
\description{
Give Studies their numbers
}
\details{
This function will add the study number to the name. This was done to match epi datasets to gwas datasets (which have them included already)
}
\examples{
GiveStudiesNumbers("whi")
}
\seealso{
\link{MakeDataAvailabilityTable} \link{CreateSurvivalDataset}
}

