% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{GrepForEpiVars}
\alias{GrepForEpiVars}
\title{GrepForEpiVars}
\usage{
GrepForEpiVars(grepTerms)
}
\arguments{
\item{grepTerms}{Which keywords to search (ex: "smoke", "BMI")}
}
\value{
Will return a vector of matching terms
}
\description{
Grep for Available Epi Variables
}
\details{
This function will search the list of variables file for matched terms. Matches are returned and you can choose
which variables to include this way.
}
\examples{
GrepForEpiVars("smoke")
GrepForEpiVars(c("BMI", "race"))
}
\seealso{
\link{MakeDataAvailabilityTable}
}

