% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{MakeDataAvailabilityTable}
\alias{MakeDataAvailabilityTable}
\title{MakeDataAvailabilityTable}
\usage{
MakeDataAvailabilityTable(epivars = NULL, studies = NULL,
  includeSurvival = FALSE, survvars = NULL)
}
\arguments{
\item{epivars}{Which covariates to include}

\item{studies}{Which studies to pull, if left NULL it will use all}

\item{includeSurvival}{Boolean, Do you want to include survival variables}

\item{survvars}{Which survival variables to include}
}
\value{
Will return a vector of matching terms
}
\description{
Make Data Availability Table for Epi Variables
}
\details{
This function will return a table with the variables of interest as columns
}
\examples{
MakeDataAvailabilityTable("smoke")
MakeDataAvailabilityTable(c("smoke", "race", "BMI"))
}
\seealso{
\link{MakeDataAvailabilityTable}
}

