% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{MakeEpiDetailsTable}
\alias{MakeEpiDetailsTable}
\title{MakeEpiDetailsTable}
\usage{
MakeEpiDetailsTable(variables)
}
\arguments{
\item{variables}{A vector of epi variables}
}
\description{
Create a Epi-Variable Table with details
}
\details{
Returns a table with the epi variable, a short description, and the permissible values.
This table can be used when we release the data.
}
\examples{
MakeEpiDetailsTable(c("age_ref", "BMI", "alcoholc"))
}
\seealso{
\link{CreateEpiDataset}
}

