% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{MakePathtoNewcomb_P}
\alias{MakePathtoNewcomb_P}
\title{Make a directory path}
\usage{
MakePathtoNewcomb_P(directory, server = "cs")
}
\arguments{
\item{directory}{The endpoint directory}

\item{server}{Which server to use, either "home" or "cs"}
}
\value{
Returns a path to the directory. This function will auto detect which host a user is on, and create a path depending on the host.
}
\description{
This function will add the root directory path to Newcomb_P folder.
}
\examples{
MakePathtoNewcomb_P("/Molecular Correlates_ISACC", server="cs")
}
\seealso{
\link{MakePathtoPeters_U}
}

