% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{MakePathtoPeters_U}
\alias{MakePathtoPeters_U}
\title{Make a directory path}
\usage{
MakePathtoPeters_U(directory, server = "home")
}
\arguments{
\item{directory}{The endpoint directory}

\item{server}{Which server to use, either "home" or "cs"}
}
\value{
Returns a path to the directory. This function will auto detect which host a user is on, and create a path depending on the host.
}
\description{
This function will add the root directory path to Peters_U folder.
}
\examples{
MakePathtoPeters_U("/GECCO_Working", server="cs")
}
\seealso{
\link{MakePathtoNewcomb_P}
}

