% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{MakePriorityPrunerInputFile}
\alias{MakePriorityPrunerInputFile}
\title{Make PriorityPruner Input File}
\usage{
MakePriorityPrunerInputFile(snplist, pvals = "gigsv2", forceSelect = NULL,
  designScore = NULL, chatty = TRUE, save.file = TRUE)
}
\arguments{
\item{snplist}{A SNP name, either rs number or chr:pos (ex, "rs1234" or "1:1234")}

\item{pvals}{Where to draw pvalues (or other values between 0 and 1). PriorityPruner will use these as a prioritization metric. This is used for prioritizing the selection of SNPs, where lower numbers are prioritized.}

\item{forceSelect}{For now, this doesn't really work other than to assign all values the same 0 status. Flag indicating if the SNP should be selected (kept) regardless of its LD with other selected SNPs or other filtering criteria specified, such as MAF or design score (1=true, 0=false).}

\item{designScore}{For now, this doesn't really work other than to assign all values the same 1 status. Flag indicating if the SNP should be selected (kept) regardless of its LD with other selected SNPs or other filtering criteria specified, such as MAF or design score (1=true, 0=false).}

\item{chatty}{Option to print progress to screen}
}
\description{
Auto create an input file for the program PriorityPruner from a vector of snp names
}
\details{
This function returns a table with the required columns for the program PriorityPruner. These include: snp_name, chr, pos, a1, a2, P, forceSelect, and designScore.
}
\examples{
MakePriorityPrunerInputFile(c("rs3181096", "rs3863057", "rs6666554"))
}
\seealso{
\link{MakeSNPDetailsTable_GIGS} \link{GetMarginal} \link{Run_PriorityPruner}
}

