% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{MakeSNPDetailsTable_GIGS}
\alias{MakeSNPDetailsTable_GIGS}
\title{MakeSNPDetailsTable_GIGS}
\usage{
MakeSNPDetailsTable_GIGS(snps, chatty = TRUE)
}
\arguments{
\item{snps}{A vector of snps (ie, c("1:123", "2:234"))}

\item{chatty}{Option to print progress to screen}
}
\description{
Create a SNP Table with details on gigs position and alleles
}
\details{
For a vector of snps, this function will pull all of the information we want to release with hapmap data.
This will create a table to use in the data release. Results will include: snp name, the count allele,
the baseline allele, R2, and CAF.
}
\examples{
MakeSNPDetailsTable_GIGS(c("5:1286516", "5:1322087", "5:1279790"))
}
\seealso{
\link{CreateSNPDetailsTable}
}

