% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{MakeSplitPos}
\alias{MakeSplitPos}
\title{MakeSplitPos}
\usage{
MakeSplitPos(gigs_numbers)
}
\arguments{
\item{gigs_numbers}{SNP names (ex, 1:132423)}
}
\description{
Internal function for splitting a snp name in gigs to its position information
}
\details{
Changes snp names from 1:123 to a data frame with name, chromosome, and position data
}
\examples{
ChangeAlleles("1")
}
\seealso{
\link{GetMarginal}
}

