% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{MergeEpiAndSurvivalData}
\alias{MergeEpiAndSurvivalData}
\title{MergeEpiAndSurvivalData}
\usage{
MergeEpiAndSurvivalData(EpiDataset, SurvivalDataset, merge_by = "compassid",
  all = TRUE)
}
\arguments{
\item{EpiDataset}{An Epi dataset}

\item{SurvivalDataset}{A Survival dataset}

\item{merge_by}{How to merge the two.  For Epi and Survival, it will always be by compassid}

\item{all}{Keep all records (TRUE) or just those that intersect (FALSE)}
}
\value{
Returns a dataset of subjects as rows, and compassID, netcdfID, study name, and variables as columns
}
\description{
Merge Epi and Survival Datasets
}
\details{
This is very straightforward function to merge two datasets by some variable.  Really could
just use the merge function, but this catches a few extra things.
}
\examples{
epi <- CreateEpiDataset(c("smoke", "BMI"))
surv <- CreateSurvivalDataset("all")
MergeEpiAndSurvivalData(epi, surv)
MergeEpiAndSurvivalData(epi, surv, all=FALSE)
}
\seealso{
\link{CreateEpiDatase} \link{CreateSurvivalDataset}
}

