% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{print.snp_location_info}
\alias{print.snp_location_info}
\title{print.snp_location_info}
\usage{
\method{print}{snp_location_info}(x, ...)
}
\arguments{
\item{x}{an object in the class "snp_location_info"}

\item{...}{Additional arguments passed to print}
}
\description{
Compact Display of SNPs locations
}
\details{
This function prints a short summary of the SNP locations to console, so it doesn't freeze up the machine if the list is huge.
}
\examples{
FindSNPposition_hapmap("rs2965667")
}
\seealso{
\link{FindSNPposition_hapmap} \link{FindSNPpositions_gigs} \link{MakeSNPDetailsTable_GIGS}
}

