% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{print.snp_names}
\alias{print.snp_names}
\title{print.snp_names}
\usage{
\method{print}{snp_names}(x, ...)
}
\arguments{
\item{x}{an object in the class "snp_names"}

\item{...}{Additional arguments passed to print}
}
\value{
This function prints a short summary of the SNP names to console, so it doesn't freeze up the machine if the list is huge.
}
\description{
Compact Display of SNP Names
}
\examples{
load(DataLocation("hapmap_snp_names"), snp_names_113nhs_omni1.Rdata")
print(snp_names)
}
\seealso{
\link{find_rs_hapmap} \link{FindSNPposition_hapmap}
}

