% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{print.snps_in_gene_regions}
\alias{print.snps_in_gene_regions}
\title{print.snps_in_gene_regions}
\usage{
\method{print}{snps_in_gene_regions}(x, ...)
}
\arguments{
\item{x}{an object in the class "snps_in_gene_regions"}

\item{...}{Additional arguments passed to print}
}
\description{
Compact Display of SNPs in gene regions
}
\details{
This function prints a short summary of the SNP names to console, so it doesn't freeze up the machine if the list is huge.
}
\examples{
Use_snp_finder.py("MYC")
}
\seealso{
\link{Use_snp_finder.py} \link{FindSNPpositions_gigs}
}

