% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{Read_PriorityPruner_Results}
\alias{Read_PriorityPruner_Results}
\title{Retrieve PriorityPruner Results}
\usage{
Read_PriorityPruner_Results(results_files)
}
\arguments{
\item{results_files}{A vector of *.results files that are output from PriorityPruner}
}
\description{
Read in results from PriorityPruner
}
\details{
This function will read in and combine all the results from a PriorityPruner analysis
}
\examples{
Read_PriorityPruner_Results(list.files(pattern="\\\\.results"))
}
\seealso{
\link{MakePriorityPrunerInputFile} \link{Run_PriorityPruner}
}

