% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{Run_PriorityPruner}
\alias{Run_PriorityPruner}
\title{Run PriorityPruner On Gizmo}
\usage{
Run_PriorityPruner(snp_table_file, datasource = "gigsv2", r2 = 0.5,
  report.call = FALSE)
}
\arguments{
\item{snp_table_file}{A snp_table file that was created either by hand or using the MakePriorityPrunerInputFile function}

\item{datasource}{This is where we draw the tfam/tped files for PriorityPruner. At this point, it can either be gigsv2 or 1kgp data}

\item{r2}{numeric, Fixed LD R2 cut off from 0-1}

\item{report.call}{Option to report the system call so you can run PriorityPruner program via terminal}
}
\description{
Run PriorityPruner on Gizmo Cluster
}
\details{
This function will batch PriorityPruner jobs to Gizmo cluster from a rhino node
}
\examples{
Run_PriorityPruner("snp_table_chr1")
}
\seealso{
\link{MakePriorityPrunerInputFile} \link{Read_PriorityPruner_Results}
}

