% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{ToKeepOrNot}
\alias{ToKeepOrNot}
\title{ToKeepOrNot}
\usage{
ToKeepOrNot(vectorofstudiestokeep)
}
\arguments{
\item{vectorofstudiestokeep}{Names vector of studies and their amount of missing data}
}
\value{
Returns boolean vector of to keep study (TRUE) or not (FALSE)
}
\description{
To Keep a Study or Discard it
}
\details{
For a vector of studies, it returns T/F to keep each or not. Really odd function I wrote...might
be worthless.  But it tries to maximize the amount of data. It will pull in studies with the same
name prefix and see which has more data and keep that one.
}
\seealso{
\link{CreateSurvivalDataset}
}

