% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{Use_snp_finder.py}
\alias{Use_snp_finder.py}
\title{Use snp_finder.py program}
\usage{
Use_snp_finder.py(gene, upstream = 0, downstream = 0, snp_list = "gigs",
  buildver = "hg19", include_ncrna = FALSE, report.call = FALSE,
  chatty = TRUE)
}
\arguments{
\item{gene}{Gene in which to find snps}

\item{upstream}{Distance upstream in bp}

\item{downstream}{Distance downstream in bp}

\item{snp_list}{Name of snp list to use. Options include: "gigs", "exome_pooled_20130624", "hapmap_imputed_20120208", or "combined_exome_1kgp"}

\item{buildver}{Which human genome build to use, either "hg18" or "hg19" (use hg18 for hapmap imputed)}

\item{include_ncrna}{Boolean, whether to include non-coding RNAs in the search (such as miRNA). Defaults to FALSE, however, if you do not find anything with that then try to run using TRUE to expand the search.}

\item{report.call}{Option to report the system call so you can run Chucks program via terminal.}

\item{chatty}{Option to print progress to screen}
}
\value{
This function uses system calls, so it will not work on a Windows machine. If the program works and finds snps, then it will return a matrix in the class "snps_in_gene_regions ", that has two columns, one with the snp location and the other with the gene.
}
\description{
Use Chuck's snp_finder.py program
}
\details{
This function is just a wrapper for Chucks snp_finder py script.  Pass it the same variables and it returns a matrix of matches.
}
\examples{
Use_snp_finder.py("MYC", 500, 500)
}
\seealso{
\link{print.snps_in_gene_regions}
}

