% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{WhichEpiFilesToInclude}
\alias{WhichEpiFilesToInclude}
\title{WhichEpiFilesToInclude}
\usage{
WhichEpiFilesToInclude(studies, files = "1or2")
}
\arguments{
\item{studies}{A vector of studies to be included in the count}

\item{files}{Argument to choose files with the extension 1/2. This is a super sucky function, because it ignores the other file types.}
}
\value{
Will return a list with two vectors. The first ($files) is a vector of file names, and the second ($sst) is a vector of the studies
}
\description{
Sift through available Epi Files
}
\details{
This function was written before chatting with Yi about the various datasets. So use with caution.
}
\examples{
WhichEpiFilesToInclude(c("101ccfr", "108dachs"))
}
\seealso{
\link{MakeDataAvailabilityTable} \link{CreateEpiDatasetPerStudy}
}

