% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SearchFunctions.R
\name{Yi_GetEpiDataFromGigs}
\alias{Yi_GetEpiDataFromGigs}
\title{Yi_GetEpiDataFromGigs}
\usage{
Yi_GetEpiDataFromGigs(env)
}
\arguments{
\item{env}{Epi variables to include}
}
\value{
Returns a dataset of subjects as rows, and variables plus lots of other info (like drop
status) as columns. Also includes PCs from gigs data.
}
\description{
Yis Function for Creating Epi Datast Using Gigs PCA
}
\examples{
env <- c('famhx1','famhx_reln1','ibd','study_site','sex','asp_ref','aspirin','cancer_site_sum1',
         'cancer_site_sum2','age_dx','BMI5','age_dxsel','stage','stage2','stage3')
Yi_GetEpiDataFromGigs(c('famhx1','ibd'))
}
\seealso{
\link{MergeEpiAndSurvivalData}
}

